+++ src/index.ts
@@ -1,5 +1,6 @@
+/** @module @foxglove/rosbag */
+
 import Bag from "./Bag";
-import BagReader from "./BagReader";
 
 export type { Filelike } from "./types";
-export { Bag, BagReader };
+export { Bag };