+++ src/node/index.ts
@@ -1,0 +1,5 @@
+/** @module @foxglove/rosbag/node */
+
+import FileReader from "./FileReader";
+
+export { FileReader };