+++ src/web/index.ts
@@ -1,0 +1,5 @@
+/** @module @foxglove/rosbag/web */
+
+import BlobReader from "./BlobReader";
+
+export { BlobReader };